[Modify]修正助教對第三週作業的回饋
